requirements.txt
asgiref==3.4.1
autopep8==1.6.0
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.4
Django==4.0.1
django-allauth==0.51.0
django-ckeditor==6.2.0
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.2.2
django-filter==21.1
django-js-asset==1.2.2
django-oauth-toolkit==2.0.0
django-vite==1.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-social-oauth2==1.2.1
ecdsa==0.17.0
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
Faker==12.2.0
google-auth==2.8.0
idna==3.3
importlib-metadata==4.10.0
jwcrypto==1.3.1
Markdown==3.3.6
oauthlib==3.2.0
Pillow==9.0.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
wrapt==1.14.1
zipp==3.7.0
